functions/requirements.txt
azure-functions
azure-identity
azure-mgmt-compute
azure-mgmt-network
requests
